Re-mix: add Rendimento and TotaldePecas *quantidade -> Kg (Double) *historico(all) to fit new changes in main table(material)

+++ src/main/java/com/projectmg/Dtos/HistoricoMaterialDTO.java
@@ -11,7 +11,7 @@ public record HistoricoMaterialDTO(
         Long id,
         String nome,
         TipoAcao acao,
-        Integer quantidadeAlterada,
+        Double quantidadeAlterada,
         String usuarioUltimaAtualizacao,
         String comentario,
         LocalDateTime dataCriacao,
